skeleton created - debounce, calculate, broadcast in widget updater class
